+++ tests/mbTests.h
@@ -4,5 +4,4 @@
 #define BOOST_TEST_MODULE mb
 #include <boost/test/unit_test.hpp>
 
-
 #endif /* __MB_TESTS_H__ */